Added vsync Changed variable types for apples Changed upgrade logic

+++ click!/Apple.cpp
@@ -2,7 +2,7 @@
 
 void Apple::calculateApplesPerSecond()
 {
-	uint64_t applesPerSecond = 0;
+	double applesPerSecond = 0;
 	for (std::vector<Upgrade>::iterator it = m_pUpgrades.begin(); it != m_pUpgrades.end(); ++it) {
 		applesPerSecond += it->getApplesPerSecond();
 	}
